doing tests and about to norm

+++ headers/ft_utils.h
@@ -9,7 +9,7 @@
 int		incharset(const char *set, char c);
 void	ft_putunsigned_fd(size_t n, int fd);
 void	ft_putunsignedbase_fd(size_t n, char *base, size_t len, int fd);
-size_t	ft_leninbase(size_t i, char *base);
-int		ft_unsignedlen(size_t n);
+size_t	ft_leninbase(unsigned int i, char *base);
+int		ft_unsignedlen(unsigned int n);
 
 #endif
